tools/sam-ba: add command line option to limit the size of applet buffers
Add a new "-L, --applet-buffer-limit" command line option to set a limit
to applet buffer sizes.

The actual applet buffer size used by sam-ba is the minimum between this
limit and the size returned by the applet during its 'initialize' command.

The former hard-coded limit of 128KB is kept as the default value for
this new option.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@microchip.com>